+++ Source/Encoding/NVEnc.vb
@@ -111,8 +111,6 @@ Public Class NVEnc
             proc.Arguments = "/S /C """ + Params.GetCommandLine(True, True) + """"
             proc.Start()
         End Using
-
-        AfterEncoding()
     End Sub
 
     Overrides Function GetMenu() As MenuList