+++ Source/Encoding/VCEEnc.vb
@@ -100,8 +100,6 @@ Public Class VCEEnc
             proc.Arguments = "/S /C """ + Params.GetCommandLine(True, True) + """"
             proc.Start()
         End Using
-
-        AfterEncoding()
     End Sub
 
     Overrides Function GetMenu() As MenuList