+++ Changelog.md
@@ -18,6 +18,7 @@ v2.35.0 (not published yet)
 ====================
 
 - UI: Set "Output Mod" default value to 4
+- UI: Fix setting Crop from/to Dolby Vision RPU files
 - Update tools
     - ...
 - Update AviSynth+ plugins